requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
langchain==0.0.184
langchain-google-genai==0.0.3
faiss-cpu==1.7.2
google-generativeai==0.1.0
python-dotenv==0.19.0
PyPDF2==2.10.5
python-docx==0.8.11
pandas==1.3.3
Pillow==8.3.2
pytesseract==0.3.8
streamlit==1.2.0
requests==2.26.0
scikit-learn==0.24.2
numpy==1.21.2
pydantic==1.8.2
starlette==0.14.2
typing-extensions==3.10.0.2